feat: 重構 SettingsModal.vue 並新增贊助頁面
