+++ frontend/components.d.ts
@@ -54,7 +54,10 @@ declare module 'vue' {
     RoomManager: typeof import('./src/components/RoomManager.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    SettingsChangelog: typeof import('./src/components/SettingsChangelog.vue')['default']
     SettingsModal: typeof import('./src/components/SettingsModal.vue')['default']
+    SettingsPreferences: typeof import('./src/components/SettingsPreferences.vue')['default']
+    SettingsSupport: typeof import('./src/components/SettingsSupport.vue')['default']
     StatusBadge: typeof import('./src/components/StatusBadge.vue')['default']
   }
 }
